+++ src/main/kotlin/com/relogging/server/global/security/jwt/service/JwtService.kt
@@ -1,0 +1,6 @@
+package com.relogging.server.global.security.jwt.service
+
+import org.springframework.stereotype.Service
+
+@Service
+class JwtService
